Dev/volume retype (#180)
* :bug: IF-15594 Fix service type for Volume class (#179)
