+++ ecl/compute/v2/volume.py
@@ -2,12 +2,11 @@
 
 from ecl.compute import compute_service
 from ecl import resource2
-from ecl import utils
 
 class Volume(resource2.Resource):
     resource_key = 'volume'
     resources_key = 'volumes'
-    base_path = '/volumes'
+    base_path = '/os-volumes'
     service = compute_service.ComputeService()
 
     # capabilities
@@ -29,28 +28,6 @@ class Volume(resource2.Resource):
     created_at = resource2.Body("createdAt")
     status = resource2.Body('status')
     bootable = resource2.Body('bootable')
-
-    def retype(self, session, volume_type, is_dry_run=False):
-        """ Volume type change and dry run configuration process. """
-        body = {
-            'os-retype': {
-                'new_type': volume_type,
-                'migration_policy': 'on-demand'
-            }
-        }
-        if is_dry_run:
-            body['os-retype']['dryRun'] = True
-        self._action(session, body)
-
-    def _action(self, session, body):
-        """Preform server actions given the message body."""
-        # NOTE: This is using Server.base_path instead of self.base_path
-        # as both Server and ServerDetail instances can be acted on, but
-        # the URL used is sans any additional /detail/ part.
-        url = utils.urljoin(Volume.base_path, self.id, 'action')
-        headers = {'Accept': ''}
-        return session.post(
-            url, endpoint_filter=self.service, json=body, headers=headers)
 
 
 class VolumeDetail(Volume):
